feat: enhance examples with interactive menu and comprehensive documentation
- Add interactive menu system to run both EmbeddedStorageExample and MonitoringExample
- Create SharedDomainClasses.cs to resolve duplicate class conflicts
- Refactor examples to use shared domain classes and eliminate code duplication
- Add comprehensive README.md documentation for examples directory
- Update project configuration to include all example files
- Implement menu-driven Program.cs with options to run individual or both examples
- Remove duplicate domain classes from individual example files
- Add cross-platform support and generic setup instructions

Changes:
- examples/ConsoleApp/Program.cs: Interactive menu system with error handling
- examples/ConsoleApp/NebulaStore.Examples.ConsoleApp.csproj: Include all example files
- examples/SharedDomainClasses.cs: Common domain classes and custom type handlers
- examples/EmbeddedStorageExample.cs: Remove duplicate classes, rename Main to RunExample
- examples/MonitoringExample.cs: Remove duplicate classes, rename Main to RunExample
- examples/README.md: Comprehensive documentation with setup, usage, and examples
